ingest_service/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
python-multipart

# Extractors (เลือกใช้ / ติดตามทีหลัง)
pymupdf
python-docx
python-pptx
Pillow
pytesseract
